backend/requirements.txt
# FastAPI Core

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client

httpx==0.26.0

# Email & Environment

python-dotenv==1.0.0
emails==0.6

# Utilities

python-dateutil==2.8.2
numpy<2.0.0

# Machine Learning / Computer Vision

torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
timm==0.9.12
opencv-python==4.9.0.72 
# opencv-python-headless (lighter and faster to install, without GUI component)
ultralytics==8.3.234
numpy==1.24.3
